+++ src/HTTP/Utils/HandleFileList.cpp
@@ -64,7 +64,7 @@ auto filterFiles(const std::vector<sFile> &files, const std::string &filePath, b
 
 // NOLINTNEXTLINE(readability-function-cognitive-complexity,misc-no-recursion)
 void handleFileList(
-        const std::shared_ptr<Cluster>& cluster, const std::string& sBundle, uint32_t jobId, bool bRecursive, const std::string &filePath,
+        const std::shared_ptr<Cluster>& cluster, const std::string& sBundle, uint64_t jobId, bool bRecursive, const std::string &filePath,
         const std::shared_ptr<HttpServerImpl::Response> &response
 ) {
     // Create a database connection
@@ -309,7 +309,7 @@ void handleFileList(
 
 void handleFileList(
         // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-        const std::shared_ptr<ClusterManager>& clusterManager, uint32_t jobId, bool bRecursive, const std::string &filePath,
+        const std::shared_ptr<ClusterManager>& clusterManager, uint64_t jobId, bool bRecursive, const std::string &filePath,
         const std::string &appName, const std::vector<std::string> &applications, const std::shared_ptr<HttpServerImpl::Response> &response
 ) {
     // Create a database connection
